+++ objectbox/lib/objectbox.dart
@@ -6,6 +6,7 @@ library objectbox;
 
 export 'src/annotations.dart';
 export 'src/box.dart' show Box, PutMode;
+export 'src/browser.dart' show Browser;
 export 'src/common.dart';
 export 'src/query.dart'
     show