+++ objectbox/lib/src/browser.dart
@@ -1,0 +1,1 @@
+export 'native/browser.dart' if (dart.library.html) 'web/browser.dart';